resolve some issue

+++ src/models/cloudinary.js
@@ -1,14 +1,15 @@
-import {v2} from "cloudinary"
+import {v2 as cloudinary} from "cloudinary"
 import fs from "fs"
 
     // Configuration
     cloudinary.config({ 
-        cloud_name: CLOUDINARY_CLOUD_NAME, 
-        api_key: CLOUDINARY_API_KEY, 
-        api_secret: CLOUDINARY_API_SECRET // Click 'View API Keys' above to copy your API secret
+        cloud_name: process.env.CLOUDINARY_CLOUD_NAME, 
+        api_key: process.env.CLOUDINARY_API_KEY, 
+        api_secret: process.env.CLOUDINARY_API_SECRET // Click 'View API Keys' above to copy your API secret
     });
 
  const uploadOnCloudinary=async(localFilePath)=>{
+    try{
         if(!localFilePath) return null
         // upload the file on cloudinary
         const response=await cloudinary.uploader.upload
@@ -16,13 +17,23 @@ import fs from "fs"
             resource_type:"auto"
         })
         //file has been upload successfull
-        console.log("file is uploaded on cloudinary",
-            response.url);
-            return response;
+      // console.log("file is uploaded on cloudinary",
+     // response.secure_url);
+       // DELETE FILE
+if(fs.existsSync(localFilePath)){
+    fs.unlinkSync(localFilePath);
+}
+   return response
+          
+      
 
     }
     catch(error){
-        fs.unlinkSync(localFilePath) // remove the locally saved temprary file as the upload operation
+        console.log("CLOUDINARY ERROR:",error);
+        if(fs.existsSync(localFilePath)){
+       fs.unlinkSync(localFilePath);
+        }
+         // remove the locally saved temprary file as the upload operation
         //got failed
         return null;
     }
